Move code around for better understandability (#85)
- No alterations have been made to the code; only documentation has been
enhanced and typos corrected.

Change Details:
* The `NodeFunctions` have been relocated to `compiler_pass.jl`.
* This transition aims to consolidate common functions and the
`CollectVariable` pass with `NodeFunctions`, thereby fostering a deeper
understanding of the code associated with `NodeFunctions`, even though
it results in a lengthier file.
* The contents of `bugsast.jl` have been transferred to `parser.jl`.
* The renaming of `@bugs` made the `bugsast` title somewhat ambiguous.
Moreover, due to the similar functionalities of `parser.jl` and
`@bugsast`, I think their merging makes sense.
* The sections of code related to the `Model` have been isolated in a
distinct file.
* This separation promotes a more streamlined abstraction of the
`Graph`. By doing this, the `Graph` interface can potentially be reused,
possibly paving the way for the development of graph-based inference
algorithms that are independent of the `Model` definitions in BUGS.
* Several functions have been migrated to `utils.jl` to minimize the
scattering of code and maintain a cleaner code base.