Remove link functions and refactor tests (#86)
* Link function field is removed from `NodeInfo`
* This only affects link function in stochastic assignments, not logical
assignments. Because in the latter, we already transformed the
expression by calling inverse function of the link function on the RHS
* `Nimble` supports this syntax while the original BUGS doesn't, so this
is not really a loss of feature
* Also with `Bijectors.jl`, the benefits of supporting link functions in
stochastic assignments is diminishing. In the future, we can add support
for specifying the `bijector` to use, which would be a solid improvement
over the link function syntax in my opinion
* The logp computation is refactored a little bit
* The way we handled link function before is to first push `dist`
through a bijector, and then pull it back and accumulate logabsdetjac.
Now this is removed
* Also a bugs fix: with `DynamicPPL.reconstruct`, matrix values are
handled correctly
* Trans-dimensional bijectors are not handled in this PR, but will be
fixed very soon
* Some tests are refactored
* Previously, we rely on overloading behavior of `DynamicPPL`, now we
just use `link` to transform the varinfo; use `LogDensityFunction` when
possible