Renaming `ConstantPropagation` to `DataTransformation` (#145)
`ConstantPropagation` compute the `transformed parameters`(Stan's
terminology), i.e. the values of variables which can be computed given
the data.

`ConstantPropagation` itself is not a bad name, given that it is
descriptive and the analogy between transformed parameters and the
compiler constant propagation is also pretty clear. But I think
`DataTransformation` is better, it should reduce possible confusions
(for people doesn't know constant propagation), closer to Stan's and
BUGS terminology(Andrew calls this `data transformation`).