`CheckRepeatedAssignment` replaced `PostChecking` (#154)
The original functionality of `PostChecking` is divided into two parts:
`CheckRepeatedAssignments` and some checks performed after
`DataTransformation`.

`PostChecking` was performed after `DataTransformation`. In the case
where repeated logical assignments, the computation done in
`DataTransformation` is wasted. `CheckRepeatedAssignments` checks
potential repeated assignment before `DataTransformation`.

`DataTransformation` use three dictionaries to track assignments,
`CheckRepeatedAssignments` use one, but divided into two groups
(`Logical` and `Stochastic`). Analysing AST as the initialization of
`CheckRepeatedAssignments` enables this.

`array_bitmaps` is no longer a result of `CheckRepeatedAssignments`. It
was used to check if a variable used on the RHS is not defined by any
statement. But if RHS uses a variable (scalar or array location) not
defined , it will error anyway because it can't be find in the
environment or have value `missing`. `array_bitmaps` also can't capture
this misuse fully. The complexity is not really worth it.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>