Update semantic analysis function interface (#160)
Two changes:
* `env` is now a field of `CompilerPass`, this allows a pass to mutate
the `env` if it wants to. This sets up for improvement coming to
`DataTransformation` to get rid of `merge_with_coalescence`
* `loop_vars` is now passed to individual `analysis_statement`
functions, this sets up for improvement coming to graph building code,
which allows us to use the same `node function` for variables born from
the same statement. In the current version, `replace_constants_in_expr`
just sub-in the loop var, end up with a different node function for each
variable

Additionally, the `check_input` function is made a bit better.
The performance is the same for all purposes.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>