Remove support for `cumulative` and `density` (#169)
This PR removes support for `density` and `cumulative` functions.

`density` and `cumulative` function in OpenBUGS takes two arguments:
e.g. `density(x, y)` computes the pdf of the distribution of `x`
evaluated at `y`; `cumulative` computes the cdf.

`density` and `cumulative` are no longer supported in `MultiBUGS` (it
was in `OpenBUGS`). And their functionality is easily supported by just
using `cdf` and `pdf` functions from `Distributions.jl`.

Supporting these functions requires: node functions have access to the
distributions of random variables at run time. In principle, this is not
impossible: we can just pass related information to the node function.
This might actually be desirable as it gives some introspection ability.

But implementing these with compiled node function is not trivial. For
now, I think it is better go for simplicity.