Add warning message for transformed data (#217)
"transformed data" are variables that can be computed with data.

In JuliaBUGS, these values are computed during compilation. And these
variables will not end up in the final graph.

For standalone language like BUGS and Stan, it make sense to add these
into the model definition. But for JuliaBUGS, it makes more sense to
compute these data transformations in Julia.

This PR add some warning message for this.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>