Change `condition` syntax  (#309)
Address https://github.com/TuringLang/JuliaBUGS.jl/issues/252

The old `condition` does two things at once: it creates a `BUGSModel`
that contains a subset of the original variables and also mark variables
to be conditioned on as `observed`. This was partially motivated by
trying to make BUGS' `Gibbs` algorithm looks like Turing.jl's while make
use the advantage of the graph.

But this is ultimately a badly thought-out interface. This PR implements
`condition` and `decondition` to be like DynamicPPL's, i.e., it only
sets variables from model parameters to observations.

I also took the chance to properly refactored the code to better handle
more types of arguments and added some docstrings.