Fix generated logdensity function (#315)
This fix some issues following the file restructuring.

It's debatable `source_generation.jl` file should be moved to `model`
folder. I left it as is.

There is an error message removed. It is tricky - I think expose the raw
error to user is probably a better idea, and we can improve the error
message if it's required.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>