Use `invokelatest` in tests (because of Julia 1.12) (#404)
Julia 1.12 makes some changes to how world age is handled.

JuliaBUGS use `eval` to generate node functions and this is problematic
under 1.12.

If a user uses JuliaBUGS `compile` in REPL, there is no problem, like
before. This issue is particularly pronounced when using JuliaBUGS with
a pattern like `compile -> LogDensityProblems.logdensity` within the
same script.

It can be pretty involved fixing the problem fundamentally. For now,
this PR patches some of the tests to use `invokelatest` so they run.

Other than it's nontrivial to fix the problem, we are likely seeing a
change to the AD interface for JuliaBUGS following
https://github.com/TuringLang/JuliaBUGS.jl/pull/397.